Update Python_Backtracking.py
Added a further simple optimization to the NQueens Logic. Now the logic skips all those rows that have a single '1'. Can't do it for columns because of the way the Python data structures work.

the simple optimization did improve the speed enough for the logic to pass the HackerEarth tests.